v1.0
This update marks the addition of all of the features I wanted to get in originally. I'll continue to implement bug fixes and good suggestions. The code is probably too messy to go on to make an arbitrary sandbox/puzzle creator from this. But that's something to explore in the future, hopefully I can rest for a while now.

* Added Shadow Visions, Wild Pyromancer, Twilight's Call, Novice Engineer, Tar Creeper, Shadow Word: Pain, Shadow Word: Death, Twilight Drake, Bloodmage Thalnos
* Added Autodefense Matrix, Noble Sacrifice, Sudden Betrayal, Wandering Monster, Repentance, Eye for an Eye, Evasion
* Added 'Favourable Deck Order?'. When on, the deck is stacked so the cards most important to the combo are on top. When off, it is shuffled randomly. (Feel free to complain if you think this is not the best order.)